PN-2712 - more splits in pf-webapp
