+++ packages/pn-personafisica-webapp/src/component/Deleghe/__test__/Delegates.test.tsx
@@ -1,5 +1,5 @@
 import { act, screen } from '@testing-library/react';
-import { axe, render } from '../../../__test__/test-utils';
+import { render } from '../../../__test__/test-utils';
 import Delegates from '../Delegates';
 import { arrayOfDelegates } from '../../../redux/delegation/__test__/test.utils';
 import * as hooks from '../../../redux/hooks';
@@ -52,14 +52,6 @@ describe('Delegates Component - assuming delegates API works properly', () => {
     expect(result.container).toHaveTextContent(/marco verdi/i);
     expect(result.container).toHaveTextContent(/davide legato/i);
     expect(result.container).not.toHaveTextContent(/luca blu/i);
-  });
-
-  it.skip('is Delegates component accessible', async ()=>{
-    const mockUseAppSelector = jest.spyOn(hooks, 'useAppSelector');
-    mockUseAppSelector.mockReturnValueOnce(arrayOfDelegates);
-    const result = render(<Delegates />);
-    const results = await axe(result?.container);
-    expect(results).toHaveNoViolations();
   });
 });
