+++ packages/pn-personafisica-webapp/src/component/Deleghe/__test__/MobileDelegates.test.tsx
@@ -1,5 +1,5 @@
 import { act, screen } from '@testing-library/react';
-import { axe, render } from '../../../__test__/test-utils';
+import { render } from '../../../__test__/test-utils';
 import { arrayOfDelegates } from '../../../redux/delegation/__test__/test.utils';
 import * as hooks from '../../../redux/hooks';
 import MobileDelegates from '../MobileDelegates';
@@ -47,14 +47,6 @@ describe('MobileDelegates Component - assuming delegates API works properly', ()
     expect(result.container).toHaveTextContent(/marco verdi/i);
     expect(result.container).toHaveTextContent(/davide legato/i);
     expect(result.container).not.toHaveTextContent(/luca blu/i);
-  });
-
-  it.skip('is Mobile Delegates component accessible', async()=>{
-    const mockUseAppSelector = jest.spyOn(hooks, 'useAppSelector');
-    mockUseAppSelector.mockReturnValueOnce(arrayOfDelegates);
-    const result = render(<MobileDelegates/>);
-    const results = await axe(result?.container);
-    expect(results).toHaveNoViolations();
   });
 });
