+++ packages/pn-personafisica-webapp/src/component/Deleghe/__test__/VerificationCodeComponent.test.tsx
@@ -1,5 +1,4 @@
 import { render } from '@testing-library/react';
-import { axe } from '../../../__test__/test-utils';
 import VerificationCodeComponent from '../VerificationCodeComponent';
 
 describe('VerificationCodeComponent', () => {
@@ -21,12 +20,4 @@ describe('VerificationCodeComponent', () => {
     expect(result.baseElement).not.toHaveTextContent(/asdfgh/i);
     expect(digitsElements).toHaveLength(fiveDigits.length);
   });
-
-  it.skip('is Verification Code component accessible', async()=>{
-    const fiveDigits = '987654321';
-    const result = render(<VerificationCodeComponent code={fiveDigits} />);
-    const results = await axe(result?.container);
-    expect(results).toHaveNoViolations();
-  });
-  
 });
