+++ packages/pn-personafisica-webapp/src/component/DomicileBanner/__test__/DomicileBanner.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent } from '@testing-library/react';
-import { axe, render } from '../../../__test__/test-utils';
+import { render } from '../../../__test__/test-utils';
 import DomicileBanner from '../DomicileBanner';
 
 jest.mock('react-i18next', () => ({
@@ -40,11 +40,5 @@ describe('DomicileBanner component', () => {
     fireEvent.click(closeButton);
     const dialog = result.queryByTestId('addDomicileBanner');
     expect(dialog).toBeNull();
-  });
-
-  it.skip('is Domicile Banner component accessible', async()=>{
-    const result = render(<DomicileBanner />);
-    const results = await axe(result?.container);
-    expect(results).toHaveNoViolations();
   });
 });
